requirements-dev.txt
# Development dependencies
pytest>=7.0.0
pytest-flask>=1.2.0
flake8>=5.0.0
black>=22.0.0
mypy>=0.990
